+++ OSSIA/ossia-qt/score/qml_util.hpp
@@ -1,21 +1,22 @@
 #pragma once
 #include <limits>
 #include <ossia/editor/scenario/time_value.hpp>
+#include <QtGlobal>
 namespace ossia
 {
 namespace qt
 {
-inline int64_t infinite() { return std::numeric_limits<int64_t>::lowest(); }
+inline qint32 infinite() { return std::numeric_limits<qint32>::lowest(); }
 
-inline bool isInfinite(int64_t t) { return t == infinite(); }
+inline bool isInfinite(qint32 t) { return t == infinite(); }
 
-inline auto defaultTime(int64_t t) -> ossia::time_value {
+inline auto defaultTime(qint32 t) -> ossia::time_value {
   return t == infinite()
       ? ossia::Infinite
       : ossia::time_value{t * 1000.};
 }
 
-inline auto reverseTime(time_value t) -> int64_t {
+inline auto reverseTime(time_value t) -> qint32 {
   return t.infinite()
       ? infinite()
       : double(t) / 1000.;
